Increase fear meter
- increases the max fear every 4 generations

+++ Assets/Scripts/GameConfig.cs
@@ -11,7 +11,7 @@ public class GameConfig : MonoBehaviour
     [Header("Fear Meter")]
     public float FearDecreaseRate = 0.005f;
     public float FearIncreaseAmount = 5f;
-    public float FearMax = 20f;
+    public int FearMax = 20;
     public float FearMeterMaxIncrease = 1;
     public float FearDecreaseIncrement = 0.0005f;
 }
